refactor(web): drop deprecated jiti() call for image-host list
SonarQube (S1874): the callable jiti(id) signature is deprecated. The
image-host allowlist is just a string array, so it doesn't need TS
transpilation at config load — moved it to a plain lib/optimizable-image-
hosts.mjs that next.config.mjs static-imports and lib/image-hosts.ts
re-exports. Single source of truth preserved; no jiti in this path.

Entire-Checkpoint: a0faf2956b57
